+++ packtools/stylechecker.py
@@ -38,14 +38,24 @@ AVAILABLE_SCHEMAS = ', '.join(sorted(['@'+key
 
 
 def get_xmlvalidator(xmlpath, no_network, extra_sch):
+    """ Get an instance of ``packtools.XMLValidator``.
+
+    :param xmlpath: filesystem or URL to an XML file.
+    :param no_network: if the parser might retrieve the DTD from the internet.
+    :param extra_sch: list of paths to schematron schemas.
+    """ 
     parsed_xml = packtools.XML(xmlpath, no_network=no_network)
-    if extra_sch:
-        path_to_extra_sch = packtools.utils.resolve_schematron_filepath(extra_sch)
+    _extra_sch = list(extra_sch)
+    if _extra_sch:
+        paths = [packtools.utils.resolve_schematron_filepath(path_or_ref)
+                 for path_or_ref in _extra_sch]
+        schemas = [packtools.utils.get_schematron_from_filepath(path)
+                   for path in paths]
     else:
-        path_to_extra_sch = None
+        schemas = None
 
     return packtools.XMLValidator.parse(parsed_xml, 
-            extra_schematron=path_to_extra_sch)
+            extra_sch_schemas=schemas)
 
 
 def annotate(validator, buff, encoding=None):
@@ -164,7 +174,7 @@ def _main():
     parser.add_argument('--loglevel', default='')  # disabled by default
     parser.add_argument('--nocolors', action='store_false',
                         help='prevents the output from being colorized by ANSI escape sequences')
-    parser.add_argument('--extrasch', default=None,
+    parser.add_argument('--extrasch', action='append', default=[],
                         help='runs an extra validation using an external schematron schema. built-in schemas are available through the prefix `@`: %s.' % AVAILABLE_SCHEMAS)
     parser.add_argument('--sysinfo', action='store_true',
                         help='show program\'s installation info and exit.')